Update and rename updater.py to start.py
